pyproject.toml
[tool.poetry]
name = "sdss-lcolamps"
version = "0.1.0a0"
description = "A library to control the LCO M2 lamps."
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/lcolamps"
repository = "https://github.com/sdss/lcolamps"
documentation = "https://sdss-lcolamps.readthedocs.io/en/latest/"
keywords = ["astronomy", "software"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "lcolamps" }
]
include = [
    {path = "lcolamps/etc/*"},
]

[tool.poetry.scripts]
lcolamps = "lcolamps.__main__:lcolamps"

[tool.poetry.dependencies]
python = "^3.9,<4.0"
sdss-clu = "^1.6.2"
sdsstools = "^0.5.2"
click_default_group = "^1.2.2"

[tool.poetry.dev-dependencies]
ipython = ">=8.0.0"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
taskipy = ">=1.1.2"
autopep8 = ">=1.5.4"
Sphinx = ">=4.0.0"
sphinx-copybutton = ">=0.3.1"
sphinx-click = ">=2.3.0"
black = {version = ">=20.8b1", allow-prereleases = true}
myst-parser = ">=0.14.0"
furo = ">=2021.7.5-beta.38"

[tool.black]
line-length = 88
target-version = ['py39']
fast = true

[tool.isort]
profile = "black"
sections = ["FUTURE", "STDLIB", "TYPING", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "lcolamps"
known_typing = ["typing"]
known_sdss = ["sdsstools", "clu"]
lines_after_imports = 2

[tool.pytest.ini_options]
addopts = "--cov lcolamps --cov-report xml --cov-report html --cov-report term"

[tool.coverage.run]
branch = true
include = ["lcolamps/*"]
omit = [
    "*/__init__.py",
    "lcolamps/__main__.py",
]

[build-system]
requires = ["poetry-core>=1.1.0b3"]
build-backend = "poetry.core.masonry.api"
